fix: fix keyboard container
